Add new fileds in event model and extract address model
