+++ backend/models/institution.py
@@ -7,6 +7,7 @@ from custom_exceptions.fieldException import FieldException
 from search_module.search_institution import SearchInstitution
 from service_messages import send_message_notification
 from service_messages import send_message_email
+from models.address import Address
 
 import json
 
@@ -16,48 +17,6 @@ def get_actuation_area(data):
     if data.get('actuation_area') == 'other':
         return data.get('other_area')
     return data.get('actuation_area')
-
-
-class Address(ndb.Model):
-    """Address model."""
-
-    number = ndb.StringProperty()
-
-    street = ndb.StringProperty()
-
-    neighbourhood = ndb.StringProperty()
-
-    city = ndb.StringProperty()
-
-    federal_state = ndb.StringProperty()
-
-    cep = ndb.StringProperty()
-
-    country = ndb.StringProperty()
-
-    def __iter__(self):
-        """Make this object iterable."""
-        yield 'number', self.number
-        yield 'street', self.street
-        yield 'neighbourhood', self.neighbourhood
-        yield 'city', self.city
-        yield 'federal_state', self.federal_state
-        yield 'cep', self.cep
-        yield 'country', self.country
-
-    @staticmethod
-    def create(data):
-        """Create an address model instance."""
-        address = Address()
-        address.number = data.get('number')
-        address.street = data.get('street')
-        address.neighbourhood = data.get('neighbourhood')
-        address.city = data.get('city')
-        address.federal_state = data.get('federal_state')
-        address.cep = data.get('cep')
-        address.country = data.get('country')
-
-        return address
 
 
 class Institution(ndb.Model):
